requirements.txt
numpy
scikit-learn
tencentcloud-sdk-python
